feature: Refactor component used frequently
