+++ client/src/App.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import "./App.css";
 import { Routes, Route } from "react-router-dom";
 import HomePage from "./components/pages/HomePage";
