+++ client/src/components/Table/TableHeader/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./TableHeader";
