+++ client/src/components/Table/TransactionTable/TransactionTableRow/TransactionTableRow.scss
@@ -1,0 +1,7 @@
+.transaction-table-row {
+  .detail {
+    color: black;
+    font-weight: lighter;
+    font-size: smaller;
+  }
+}
