--- client/src/components/TransactionTable/TransactionTableHeader/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./TransactionTableHeader";
